libsyn123: silence some warnings
git-svn-id: svn://scm.orgis.org/mpg123/trunk@5287 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libsyn123/geiger.c
@@ -224,7 +224,7 @@ static void geiger_init(struct geigerspace *gs, double activity, long rate)
 	// Experimenting, actually. Some relation to the speaker.
 	gs->force_scale = 50000.*gs->mass*0.001/(4.*gs->dead_s*gs->dead_s);
 
-	float event_likelihood = activity*gs->time_interval;
+	float event_likelihood = (float)(activity*gs->time_interval);
 	if(event_likelihood > 1.f)
 		event_likelihood = 1.f;
 	gs->thres = 1.f-event_likelihood;
